mitosheet: cleanup code by creating helper function
